requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
Naked==0.1.32
pycryptodome==3.23.0
PyYAML==6.0.3
requests==2.32.5
shellescape==3.8.1
urllib3==2.6.3
